+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/sta/model/MultiDatastream.java
@@ -262,7 +262,7 @@ public class MultiDatastream extends AbstractEntity {
 
     public void setName(String name) {
         this.name = name;
-        setName = true;
+        setName = name != null;
     }
 
     /**
@@ -270,7 +270,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setDescription(String description) {
         this.description = description;
-        setDescription = true;
+        setDescription = description != null;
     }
 
     /**
@@ -278,7 +278,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setObservationType(String observationType) {
         this.observationType = observationType;
-        setObservationType = true;
+        setObservationType = observationType != null;
     }
 
     /**
@@ -286,7 +286,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setMultiObservationDataTypes(List<String> observationTypes) {
         this.multiObservationDataTypes = observationTypes;
-        setMultiObservationDataTypes = true;
+        setMultiObservationDataTypes = multiObservationDataTypes != null;
     }
 
     /**
@@ -294,7 +294,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setUnitOfMeasurements(List<UnitOfMeasurement> unitsOfMeasurement) {
         this.unitOfMeasurements = unitsOfMeasurement;
-        setUnitOfMeasurements = true;
+        setUnitOfMeasurements = unitOfMeasurements != null;
     }
 
     /**
@@ -302,7 +302,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setSensor(Sensor sensor) {
         this.sensor = sensor;
-        setSensor = true;
+        setSensor = sensor != null;
     }
 
     /**
@@ -310,7 +310,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setObservedProperties(EntitySet<ObservedProperty> observedProperties) {
         this.observedProperties = observedProperties;
-        setObservedProperties = true;
+        setObservedProperties = observedProperties != null;
     }
 
     /**
@@ -325,7 +325,7 @@ public class MultiDatastream extends AbstractEntity {
      */
     public void setThing(Thing thing) {
         this.thing = thing;
-        setThing = true;
+        setThing = thing != null;
     }
 
     @Override