+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/sta/model/Thing.java
@@ -136,12 +136,12 @@ public class Thing extends AbstractEntity {
 
     public void setName(String name) {
         this.name = name;
-        setName = true;
+        setName = name != null;
     }
 
     public void setDescription(String description) {
         this.description = description;
-        setDescription = true;
+        setDescription = description != null;
     }
 
     public void setProperties(Map<String, Object> properties) {