+++ FROST-Server.Core/src/test/java/de/fraunhofer/iosb/ilt/sta/serialize/EntityFormatterTest.java
@@ -17,11 +17,11 @@
  */
 package de.fraunhofer.iosb.ilt.sta.serialize;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.fraunhofer.iosb.ilt.sta.formatter.DataArrayResult;
 import de.fraunhofer.iosb.ilt.sta.formatter.DataArrayValue;
+import de.fraunhofer.iosb.ilt.sta.json.deserialize.EntityParser;
 import de.fraunhofer.iosb.ilt.sta.json.serialize.EntityFormatter;
 import de.fraunhofer.iosb.ilt.sta.model.Actuator;
 import de.fraunhofer.iosb.ilt.sta.model.Datastream;
@@ -56,8 +56,10 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.joda.time.DateTime;
@@ -111,7 +113,7 @@ public class EntityFormatterTest {
                 .setProperties(parameters)
                 .build();
 
-        Assert.assertTrue(jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        Assert.assertTrue(jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -133,7 +135,7 @@ public class EntityFormatterTest {
                 .setTaskingParameters(parameters)
                 .build();
 
-        Assert.assertTrue(jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        Assert.assertTrue(jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -165,7 +167,7 @@ public class EntityFormatterTest {
                 .setTaskingParameters(parameters)
                 .build();
 
-        Assert.assertTrue(jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        Assert.assertTrue(jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -195,7 +197,7 @@ public class EntityFormatterTest {
                 .addProperty("owner", "John Doe")
                 .addProperty("color", "Silver")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -216,8 +218,11 @@ public class EntityFormatterTest {
                 .addProperty("owner", "John Doe")
                 .addProperty("color", "Silver")
                 .build();
-        List<Property> selectedProps = Arrays.asList(EntityProperty.Id, EntityProperty.Name);
-        assert (jsonEqual(expResult, new EntityFormatter(selectedProps).writeEntity(entity)));
+        Set<Property> selectedProps = new HashSet<>();
+        selectedProps.add(EntityProperty.Id);
+        selectedProps.add(EntityProperty.Name);
+        entity.setSelectedProperties(selectedProps);
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -255,7 +260,7 @@ public class EntityFormatterTest {
         EntitySet<Thing> things = new EntitySetImpl<>(EntityType.Thing);
         things.add(entity);
         things.add(entity);
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntityCollection(things)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntityCollection(things)));
     }
 
     @Test
@@ -264,7 +269,7 @@ public class EntityFormatterTest {
                 = "{}";
         Thing entity = new ThingBuilder()
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -338,7 +343,7 @@ public class EntityFormatterTest {
         EntitySet<Thing> things = new EntitySetImpl<>(EntityType.Thing);
         things.add(entity);
         things.setCount(1);
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntityCollection(things)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntityCollection(things)));
     }
 
     @Test
@@ -370,7 +375,56 @@ public class EntityFormatterTest {
                 .addProperty("owner", "John Doe")
                 .addProperty("color", "Silver")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
+
+        expResult
+                = "{\n"
+                + "\"@iot.id\": 1,\n"
+                + "\"Locations@iot.navigationLink\": \"Things(1)/Locations\",\n"
+                + "\"Datastreams\": [{\"@iot.id\":123}],\n"
+                + "\"name\": \"This thing is an oven.\"\n"
+                + "}";
+        entity = new ThingBuilder()
+                .setId(new IdLong(1))
+                .setSelfLink("http://example.org/v1.0/Things(1)")
+                .setLocations(new EntitySetImpl(EntityType.Location, "Things(1)/Locations"))
+                .addDatastream(new DatastreamBuilder()
+                        .setId(new IdLong(123))
+                        .build())
+                .setHistoricalLocations(new EntitySetImpl(EntityType.HistoricalLocation, "Things(1)/HistoricalLocations"))
+                .setName("This thing is an oven.")
+                .setDescription("This thing is an oven.")
+                .addProperty("owner", "John Doe")
+                .addProperty("color", "Silver")
+                .build();
+        entity.setSelectedPropertyNames(new HashSet<>(Arrays.asList(EntityProperty.Id.getJsonName(), "name", "Locations")));
+        entity.getDatastreams().setExportObject(true);
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
+
+        expResult
+                = "{\n"
+                + "\"@iot.selfLink\": \"http://example.org/v1.0/Things(1)\",\n"
+                + "\"Locations@iot.navigationLink\": \"Things(1)/Locations\",\n"
+                + "\"Datastreams\": [{\"@iot.id\":123}],\n"
+                + "\"name\": \"This thing is an oven.\"\n"
+                + "}";
+        entity = new ThingBuilder()
+                .setId(new IdLong(1))
+                .setSelfLink("http://example.org/v1.0/Things(1)")
+                .setLocations(new EntitySetImpl(EntityType.Location, "Things(1)/Locations"))
+                .addDatastream(new DatastreamBuilder()
+                        .setId(new IdLong(123))
+                        .build())
+                .setHistoricalLocations(new EntitySetImpl(EntityType.HistoricalLocation, "Things(1)/HistoricalLocations"))
+                .setName("This thing is an oven.")
+                .setDescription("This thing is an oven.")
+                .addProperty("owner", "John Doe")
+                .addProperty("color", "Silver")
+                .build();
+        entity.setSelectedPropertyNames(new HashSet<>(Arrays.asList(EntityProperty.SelfLink.getJsonName(), "name", "Locations")));
+        entity.getDatastreams().setExportObject(true);
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
+
         expResult
                 = "{\n"
                 + "  \"@iot.id\": 1,\n"
@@ -396,7 +450,7 @@ public class EntityFormatterTest {
                 .addProperty("color", "Silver")
                 .build();
         entity.getDatastreams().setExportObject(true);
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -417,7 +471,7 @@ public class EntityFormatterTest {
                     .setHistoricalLocations(new EntitySetImpl(EntityType.HistoricalLocation, "Locations(1)/HistoricalLocations"))
                     .setEncodingType("application/vnd.geo+json")
                     .build();
-            assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+            assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
         }
         {
             String expResult
@@ -435,7 +489,7 @@ public class EntityFormatterTest {
                     .setHistoricalLocations(new EntitySetImpl(EntityType.HistoricalLocation, "Locations(1)/HistoricalLocations"))
                     .setEncodingType("application/geo+json")
                     .build();
-            assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+            assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
         }
     }
 
@@ -467,7 +521,7 @@ public class EntityFormatterTest {
                 .setEncodingType("application/vnd.geo+json")
                 .setLocation(TestHelper.getFeatureWithPoint(-114.06, 51.05))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -487,7 +541,7 @@ public class EntityFormatterTest {
                 .setThing(new ThingBuilder().setNavigationLink("HistoricalLocations(1)/Thing").build())
                 .setTime(TestHelper.createTimeInstant(2015, 01, 25, 12, 0, 0, DateTimeZone.forOffsetHours(-7), DateTimeZone.UTC))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -530,7 +584,7 @@ public class EntityFormatterTest {
                 .setPhenomenonTime(TestHelper.createTimeInterval(2014, 03, 1, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .setResultTime(TestHelper.createTimeInterval(2014, 03, 01, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -568,7 +622,7 @@ public class EntityFormatterTest {
                 .setPhenomenonTime(TestHelper.createTimeInterval(2014, 03, 1, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .setResultTime(TestHelper.createTimeInterval(2014, 03, 01, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -617,7 +671,7 @@ public class EntityFormatterTest {
                 .setPhenomenonTime(TestHelper.createTimeInterval(2014, 03, 1, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .setResultTime(TestHelper.createTimeInterval(2014, 03, 01, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -676,7 +730,7 @@ public class EntityFormatterTest {
                 .setPhenomenonTime(TestHelper.createTimeInterval(2014, 03, 1, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .setResultTime(TestHelper.createTimeInterval(2014, 03, 01, 13, 0, 0, 2015, 05, 11, 15, 30, 0, DateTimeZone.UTC))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -700,7 +754,7 @@ public class EntityFormatterTest {
                 .setEncodingType("application/pdf")
                 .setMetadata("http://example.org/TMP35_36_37.pdf")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -723,7 +777,7 @@ public class EntityFormatterTest {
                 .setEncodingType("application/pdf")
                 .setMetadata("http://example.org/TMP35_36_37.pdf")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -745,7 +799,7 @@ public class EntityFormatterTest {
                 .setName("DewPoint Temperature")
                 .setDefinition("http://dbpedia.org/page/Dew_point")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -769,7 +823,31 @@ public class EntityFormatterTest {
                 .setResultTime(TestHelper.createTimeInstantUTC(2014, 12, 31, 19, 59, 59))
                 .setResult(new BigDecimal("70.40"))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
+    }
+
+    @Test
+    public void writeObservation_BasicWithNullResult() throws Exception {
+        String expResult
+                = "{\n"
+                + "	\"@iot.id\": 1,\n"
+                + "	\"@iot.selfLink\": \"http://example.org/v1.0/Observations(1)\",\n"
+                + "	\"FeatureOfInterest@iot.navigationLink\": \"Observations(1)/FeatureOfInterest\",\n"
+                + "	\"Datastream@iot.navigationLink\":\"Observations(1)/Datastream\",\n"
+                + "	\"phenomenonTime\": \"2014-12-31T11:59:59.000Z\",\n"
+                + "	\"resultTime\": \"2014-12-31T19:59:59.000Z\",\n"
+                + "	\"result\": null\n"
+                + "}";
+        Entity entity = new ObservationBuilder()
+                .setId(new IdLong(1))
+                .setSelfLink("http://example.org/v1.0/Observations(1)")
+                .setFeatureOfInterest(new FeatureOfInterestBuilder().setNavigationLink("Observations(1)/FeatureOfInterest").build())
+                .setDatastream(new DatastreamBuilder().setNavigationLink("Observations(1)/Datastream").build())
+                .setPhenomenonTime(TestHelper.createTimeInstantUTC(2014, 12, 31, 11, 59, 59))
+                .setResultTime(TestHelper.createTimeInstantUTC(2014, 12, 31, 19, 59, 59))
+                .setResult(null)
+                .build();
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -793,7 +871,7 @@ public class EntityFormatterTest {
                 .setResultTime(new TimeInstant(null))
                 .setResult("70.4")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -816,7 +894,7 @@ public class EntityFormatterTest {
                 .setResultTime(TestHelper.createTimeInstantUTC(2014, 12, 31, 19, 59, 59))
                 .setResult("70.4")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -920,7 +998,7 @@ public class EntityFormatterTest {
         source.getValue().add(dav2);
         source.getValue().add(dav3);
 
-        assert (jsonEqual(expResult, new EntityFormatter().writeObject(source)));
+        assert (jsonEqual(expResult, EntityFormatter.writeObject(source)));
     }
 
     @Test
@@ -942,7 +1020,7 @@ public class EntityFormatterTest {
                 .setDescription("This is a weather station.")
                 .setEncodingType("application/vnd.geo+json")
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     @Test
@@ -975,11 +1053,11 @@ public class EntityFormatterTest {
                 .setEncodingType("application/vnd.geo+json")
                 .setFeature(TestHelper.getFeatureWithPoint(-114.06, 51.05))
                 .build();
-        assert (jsonEqual(expResult, new EntityFormatter().writeEntity(entity)));
+        assert (jsonEqual(expResult, EntityFormatter.writeEntity(entity)));
     }
 
     private boolean jsonEqual(String string1, String string2) {
-        ObjectMapper mapper = new ObjectMapper().configure(DeserializationFeature.USE_BIG_DECIMAL_FOR_FLOATS, true);
+        ObjectMapper mapper = EntityParser.getSimpleObjectMapper();
         try {
             JsonNode json1 = mapper.readTree(string1);
             JsonNode json2 = mapper.readTree(string2);