+++ FROST-Server.SQL.PGLong/src/main/java/de/fraunhofer/iosb/ilt/sta/persistence/postgres/longid/PropertyHelper.java
@@ -19,12 +19,12 @@ package de.fraunhofer.iosb.ilt.sta.persistence.postgres.longid;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mysema.commons.lang.CloseableIterator;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.Expression;
 import com.querydsl.core.types.Path;
 import com.querydsl.core.types.dsl.NumberPath;
+import de.fraunhofer.iosb.ilt.sta.json.deserialize.EntityParser;
 import de.fraunhofer.iosb.ilt.sta.json.deserialize.custom.GeoJsonDeserializier;
 import de.fraunhofer.iosb.ilt.sta.model.Actuator;
 import de.fraunhofer.iosb.ilt.sta.model.Datastream;
@@ -867,7 +867,7 @@ public class PropertyHelper {
         if (id == null) {
             return null;
         }
-        Datastream ds = new Datastream();
+        Datastream ds = new Datastream(true);
         ds.setId(new IdLong(id));
         ds.setExportObject(false);
         return ds;
@@ -937,7 +937,7 @@ public class PropertyHelper {
         if (locationString == null || locationString.isEmpty()) {
             return null;
         }
-        if (encodingType != null && GeoJsonDeserializier.encodings.contains(encodingType.toLowerCase())) {
+        if (encodingType != null && GeoJsonDeserializier.ENCODINGS.contains(encodingType.toLowerCase())) {
             try {
                 Object geoJson = new GeoJsonDeserializier().deserialize(locationString);
                 return geoJson;
@@ -964,7 +964,7 @@ public class PropertyHelper {
         }
 
         try {
-            return new ObjectMapper().readTree(json);
+            return EntityParser.getSimpleObjectMapper().readTree(json);
         } catch (IOException ex) {
             throw new IllegalStateException("Failed to parse stored json.", ex);
         }
@@ -975,7 +975,7 @@ public class PropertyHelper {
             return null;
         }
         try {
-            return new ObjectMapper().readValue(json, clazz);
+            return EntityParser.getSimpleObjectMapper().readValue(json, clazz);
         } catch (IOException ex) {
             throw new IllegalStateException("Failed to parse stored json.", ex);
         }
@@ -986,7 +986,7 @@ public class PropertyHelper {
             return null;
         }
         try {
-            return new ObjectMapper().readValue(json, typeReference);
+            return EntityParser.getSimpleObjectMapper().readValue(json, typeReference);
         } catch (IOException ex) {
             throw new IllegalStateException("Failed to parse stored json.", ex);
         }