+++ FROST-Server.SQL.PGUuid/src/main/java/de/fraunhofer/iosb/ilt/sta/persistence/postgres/uuidid/EntityInserter.java
@@ -30,6 +30,7 @@ import com.querydsl.sql.SQLQuery;
 import com.querydsl.sql.SQLQueryFactory;
 import com.querydsl.sql.dml.SQLInsertClause;
 import com.querydsl.sql.dml.SQLUpdateClause;
+import de.fraunhofer.iosb.ilt.sta.json.deserialize.EntityParser;
 import de.fraunhofer.iosb.ilt.sta.json.deserialize.custom.GeoJsonDeserializier;
 import de.fraunhofer.iosb.ilt.sta.json.serialize.GeoJsonSerializer;
 import de.fraunhofer.iosb.ilt.sta.messagebus.EntityChangedMessage;
@@ -531,9 +532,8 @@ public class EntityInserter {
         QThings qt = QThings.things;
         QDatastreams qd = QDatastreams.datastreams;
         QMultiDatastreams qmd = QMultiDatastreams.multiDatastreams;
-        // TODO: Should probably contain a where that only returns locations
-        // with a supported encoding type.
-        SQLQuery<Tuple> query = qf.select(ql.id, ql.genFoiId)
+
+        SQLQuery<Tuple> query = qf.select(ql.id, ql.genFoiId, ql.encodingType)
                 .from(ql)
                 .innerJoin(qtl).on(ql.id.eq(qtl.locationId))
                 .innerJoin(qt).on(qt.id.eq(qtl.thingId));
@@ -544,20 +544,37 @@ public class EntityInserter {
             query.innerJoin(qd).on(qd.thingId.eq(qt.id))
                     .where(qd.id.eq(dsId));
         }
-        Tuple tuple = query.fetchOne();
-        if (tuple == null) {
+        List<Tuple> tuples = query.fetch();
+        if (tuples.isEmpty()) {
             // Can not generate foi from Thing with no locations.
             throw new NoSuchEntityException("Can not generate foi for Thing with no locations.");
         }
-        UUID genFoiId = tuple.get(ql.genFoiId);
-        UUID locationId = tuple.get(ql.id);
+        // See if any of the locations have a generated foi.
+        // Also track if any of the location has a supported encoding type.
+        UUID genFoiId = null;
+        UUID locationId = null;
+        for (Tuple tuple : tuples) {
+            genFoiId = tuple.get(ql.genFoiId);
+            if (genFoiId != null) {
+                break;
+            }
+            String encodingType = tuple.get(ql.encodingType);
+            if (encodingType != null && GeoJsonDeserializier.ENCODINGS.contains(encodingType.toLowerCase())) {
+                locationId = tuple.get(ql.id);
+            }
+        }
+        // Either genFoiId will have a value, if a generated foi was found,
+        // Or locationId will have a value if a supported encoding type was found.
 
         FeatureOfInterest foi;
-        if (genFoiId == null) {
+        if (genFoiId != null) {
+            foi = new FeatureOfInterest();
+            foi.setId(new UuidId(genFoiId));
+        } else if (locationId != null) {
             query = qf.select(ql.id, ql.encodingType, ql.location)
                     .from(ql)
                     .where(ql.id.eq(locationId));
-            tuple = query.fetchOne();
+            Tuple tuple = query.fetchOne();
             if (tuple == null) {
                 // Can not generate foi from Thing with no locations.
                 // Should not happen, since the query succeeded just before.
@@ -580,8 +597,8 @@ public class EntityInserter {
                     .execute();
             LOGGER.debug("Generated foi {} from Location {}.", foiId, locationId);
         } else {
-            foi = new FeatureOfInterest();
-            foi.setId(new UuidId(genFoiId));
+            // Can not generate foi from Thing with no locations.
+            throw new NoSuchEntityException("Can not generate foi for Thing, all locations have an un supported encoding type.");
         }
         return foi;
     }
@@ -1504,7 +1521,7 @@ public class EntityInserter {
      * @return The insert or update clause.
      */
     private <T extends StoreClause> T insertGeometry(T clause, StringPath locationPath, GeometryPath<Geometry> geomPath, String encodingType, final Object location) {
-        if (encodingType != null && GeoJsonDeserializier.encodings.contains(encodingType.toLowerCase())) {
+        if (encodingType != null && GeoJsonDeserializier.ENCODINGS.contains(encodingType.toLowerCase())) {
             String locJson;
             try {
                 locJson = new GeoJsonSerializer().serialize(location);
@@ -1709,7 +1726,7 @@ public class EntityInserter {
 
     public ObjectMapper getFormatter() {
         if (formatter == null) {
-            formatter = new ObjectMapper();
+            formatter = EntityParser.getSimpleObjectMapper();
         }
         return formatter;
     }