+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 
+# Release Version 1.7
+Version 1.7 is not released yet.
+
+**New Features**
+* Observation.result can be explicitly set to null. This is useful in cases where
+  an observation did not produce a value, but the fact that an observation was attempted
+  must still be recorded.
+* Exposed database connection options `persistence.db.conn.max`, `persistence.db.conn.idle.max`, `persistence.db.conn.idle.min`
+
+**Bugfixes**
+* Fixed #53: Query parser not Unicode aware.
+* Fixed #52: Generating FeatureOfInterest did not work for Things with multiple Location entities when some of these entities were not geoJSON.
+* Fixed the 'year' function not working on interval properties.
+
+
 # Release Version 1.6
-Version 1.6 has not been released yet.
+Version 1.6 was released on 2018-05-09.
 
 **New Features**
 * User-defined-ids. FROST-Server can not be configured to allow the user to specify the id of created enitites.
@@ -12,7 +27,7 @@ Version 1.6 has not been released yet.
   Thanks to Marcel Köpke for the patch.
 * Improved time handling in queries. FROST-Server can now calculate with times:
 
-    ```/Observations?$filter=phenomenonTime gt now() sub duration'P1D' mul (Datastream/properties/days)```
+    ```/Observations?$filter=phenomenonTime gt now() sub duration'P1D' mul Datastream/properties/days```
 
 * Separated the MQTT and HTTP parts of the server.
   The MQTT and HTTP parts of the server are now separated in to stand-alone programs: